requirements.txt
Flask==3.0.0
gunicorn==23.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.1.0
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
requests==2.31.0
google-cloud-vision==3.7.2
openai==1.54.3
google-generativeai==0.8.3
anthropic==0.39.0
